pyproject.toml
[project]
name = "emailing-ai-agent"
version = "0.1.0"
description = "Emailing AI Agent"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.49.0",
    "gradio>=5.22.0",
    "openai-agents",
    "pypdf>=6.7.4",
    "python-dotenv>=1.0.0",
    "sendgrid>=6.11.0",
]
